Move rule collection to domain
