--- src/Application/Rule/Rules/Rules.php
@@ -1,50 +1,0 @@
-<?php
-
-namespace ArtARTs36\MergeRequestLinter\Application\Rule\Rules;
-
-use ArtARTs36\MergeRequestLinter\Shared\DataStructure\Arrayee;
-use ArtARTs36\MergeRequestLinter\Domain\Rule\Rule;
-
-/**
- * @template-extends Arrayee<int, Rule>
- */
-class Rules extends Arrayee implements \ArtARTs36\MergeRequestLinter\Domain\Rule\Rules
-{
-    /**
-     * @param iterable<Rule> $rules
-     */
-    public static function make(iterable $rules): self
-    {
-        $instance = new self([]);
-
-        foreach ($rules as $rule) {
-            $instance->add($rule);
-        }
-
-        return $instance;
-    }
-
-    public function add(Rule $rule): self
-    {
-        $this->items[] = $rule;
-
-        $this->count++;
-
-        return $this;
-    }
-
-    public function implodeNames(string $sep): string
-    {
-        $str = '';
-
-        foreach ($this->items as $item) {
-            $str .= $item->getName();
-
-            if (next($this->items) !== false) {
-                $str .= $sep;
-            }
-        }
-
-        return $str;
-    }
-}
